+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: `DashboardLayout`: Add position style to `main`.
+- Fix: `DashboardLayout`: Add position style to `main`.
+- Feature: `LoaderWithParagraph`: Add `loaderComponent` prop for custom component.
 - Feature: `PlayerIconNext`: Add icon component.
 - Fix: `Video`: Update style.
 - Fix: `EmbedPlayer`: Update style.